Add a sensible default log level for test logging
Log nimiq crates with `TRACE` level and everything else with `INFO`.
This can still be customized with the `RUST_LOG` environment variable.

To do this, some logging-related code was extracted into the `nimiq-log`
crate.

Fixes #723.
