+++ lib/Cargo.toml
@@ -19,7 +19,6 @@ maintenance = { status = "experimental" }
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-ansi_term = "0.12"
 clap = { version = "3.0", features = ["derive"] }
 console-subscriber = { version = "0.1", features = ["parking_lot"], optional = true }
 derive_builder = "0.11"
@@ -39,7 +38,6 @@ toml = "0.5"
 url = { version = "2.2", features = ["serde"] }
 thiserror = "1.0"
 tokio = { version = "1.16", features = ["rt", "tracing"], optional = true }
-tracing-log = { version = "0.1", optional = true }
 tracing-loki = { version = "0.1.0", optional = true }
 tracing-subscriber = { version = "0.3", optional = true }
 
@@ -53,6 +51,7 @@ nimiq-genesis = { path = "../genesis" }
 nimiq-jsonrpc-core = { git = "https://github.com/nimiq/jsonrpc.git" }
 nimiq-jsonrpc-server = { git = "https://github.com/nimiq/jsonrpc.git" }
 nimiq-keys = { path = "../keys" }
+nimiq-log = { path = "../log", optional = true }
 nimiq-mempool = { path = "../mempool" }
 nimiq-network-libp2p = { path = "../network-libp2p" }
 nimiq-network-interface = { path = "../network-interface" }
@@ -71,7 +70,7 @@ nimiq-test-log = { path = "../test-log" }
 deadlock = []
 default = []
 launcher = []
-logging = ["console-subscriber", "serde_json", "tokio", "tracing-log", "tracing-loki", "tracing-subscriber"]
+logging = ["console-subscriber", "nimiq-log", "serde_json", "tokio", "tracing-loki", "tracing-subscriber"]
 panic = ["log-panics"]
 rpc-server = ["validator", "nimiq-rpc-server", "nimiq-wallet"]
 validator = ["nimiq-validator", "nimiq-validator-network", "nimiq-bls", "nimiq-rpc-server"]
