+++ log/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "nimiq-log"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+ansi_term = "0.12"
+log = { package = "tracing", version = "0.1", features = ["log"] }
+tracing-log = "0.1"
+tracing-subscriber = "0.3"
