+++ test-log/Cargo.toml
@@ -4,6 +4,8 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+log = { package = "tracing", version = "0.1", features = ["log"] }
+nimiq-log = { path = "../log" }
 nimiq-test-log-proc-macro = { path = "proc-macro" }
 parking_lot = { git = "https://github.com/styppo/parking_lot.git" }
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
