+++ test-log/src/lib.rs
@@ -1,4 +1,9 @@
+use log::level_filters::LevelFilter;
+use nimiq_log::TargetsExt;
 use parking_lot::Once;
+use tracing_subscriber::filter::Targets;
+use tracing_subscriber::layer::SubscriberExt;
+use tracing_subscriber::util::SubscriberInitExt;
 
 pub use nimiq_test_log_proc_macro::test;
 
@@ -7,9 +12,14 @@ static INITIALIZE: Once = Once::new();
 #[doc(hidden)]
 pub fn initialize() {
     INITIALIZE.call_once(|| {
-        tracing_subscriber::fmt()
-            .with_env_filter(tracing_subscriber::EnvFilter::from_default_env())
-            .with_test_writer()
+        tracing_subscriber::registry()
+            .with(tracing_subscriber::fmt::layer().with_test_writer())
+            .with(
+                Targets::new()
+                    .with_default(LevelFilter::INFO)
+                    .with_nimiq_targets(LevelFilter::DEBUG)
+                    .with_env(),
+            )
             .init();
     });
 }
